Update thread ID to model as long

+++ src/main/kotlin/io/github/inductiveautomation/kindling/thread/comparison/BlockerButton.kt
@@ -4,7 +4,7 @@ import com.formdev.flatlaf.extras.components.FlatButton
 import io.github.inductiveautomation.kindling.utils.FlatActionIcon
 
 internal class BlockerButton : FlatButton() {
-    var blocker: Int? = null
+    var blocker: Long? = null
         set(value) {
             isVisible = value != null
             text = value?.toString()
